One error less: now it's safe to use self upon self
